(MODULES-8232) fixes for snmp notification receiver to support unsetting
The existing code allowed you to unset a value through -1/unset and was missed during the conversion.

Changes still need to happen to netdev_stdlib to support this.

The deletion of the existing snmp receiver was missed during conversion and meant that multiple receivers with same host/name would be created causing problems.